Fix issue #3 (graphical glitches with text objects in release)
Previously agl::createCairoContext used malloc to allocate a
buffer for the pixel data, but when a buffer was allocated again
after having previously freed an old one, the existing pixel data
was retained. This commit fixes this issue by using calloc
instead, which zeroes the buffer. As a bonus, in Debug the texture
does not have a weird semi-transparent background.